Make Makefile a bit more generic
udevadm is under /bin on Gentoo, so being a bit more vague should make
it work on both Ubuntu and Gentoo. Would be nice to test on Redhat based
distros too.

Added ccflags-y to be super hard-headed about warnings, just because I
hate compiling stuff and seeing hundreds of warnings.