Change vendor ID to match Kernel
This constant needs to be added to hid-quirks in the kernel source so
HID no longer grabs the tablet. USB_DEVICE_ID_BOSTO14WA_2 because this
is generation 2 of the 14WA and I want to minimise the confusion already
caused by that. As far as I know generation 1 looks like a Wacom to the
system.